Amend alembic migration
