+++ alembic_users/versions/a8e57fe0d82e_add_columns_for_ai_quota.py
@@ -1,0 +1,40 @@
+"""Add coloumns for AI quota
+
+Revision ID: a8e57fe0d82e
+Revises: 84960b7d968c
+Create Date: 2024-09-03 18:48:00.917543
+
+"""
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.engine.reflection import Inspector
+
+
+# revision identifiers, used by Alembic.
+revision = "a8e57fe0d82e"
+down_revision = "84960b7d968c"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    conn = op.get_bind()
+    inspector = Inspector.from_engine(conn)
+    columns = [col["name"] for col in inspector.get_columns("trees")]
+    if "quota_ai" not in columns:
+        op.add_column("trees", sa.Column("quota_ai", sa.Integer(), nullable=True))
+    if "usage_ai" not in columns:
+        op.add_column("trees", sa.Column("usage_ai", sa.Integer(), nullable=True))
+    if "min_role_ai" not in columns:
+        op.add_column("trees", sa.Column("min_role_ai", sa.Integer(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column("trees", "usage_ai")
+    op.drop_column("trees", "quota_ai")
+    op.drop_column("trees", "min_role_ai")
+    # ### end Alembic commands ###
